Rename core tests for legacy multi_ptr
Separate commit provided to avoid bogus "renaming"
from the "_fp16.cpp" to the "_core.cpp"
